Create pheonix_and_balance_codeforces
Phoenix has n coins with weights 21,22,…,2n. He knows that n

is even.

He wants to split the coins into two piles such that each pile has exactly n2
coins and the difference of weights between the two piles is minimized. Formally, let a denote the sum of weights in the first pile, and b denote the sum of weights in the second pile. Help Phoenix minimize |a−b|, the absolute value of a−b

.
Input

The input consists of multiple test cases. The first line contains an integer t
(1≤t≤100

) — the number of test cases.

The first line of each test case contains an integer n
(2≤n≤30; n

is even) — the number of coins that Phoenix has.
Output

For each test case, output one integer — the minimum possible difference of weights between the two piles.