Find the Celebrity among the n persons.
We are given an adjacency matrix wherein we have 1s and 0s. 1 in (i,j) indicates that ith person knows the jth person. A celebrity is someone who knows no one but is known to everyone. We  need to find it in O(n) time and O(1) space complexity. 
The approach to use is "Two pointer approach" wherein we set our pointers to start and end and while start<end, if start knows end, then start can't be the celebrity, so move the pointer one ahead and vice versa for the end pointer. if both know each other then move both in each other's direction. If you get  one potential candidate at the end of this process, then check that kth row and kth column to see if he is known to all and he knows no one. This is done in O(n) time complexity.