time - O(n), space - O(1)
We are given an array of integers. We need to find the next permutation that is possible for it. If none is possible, then return the array sorted in ascending order. We need to find the peak which is nearest to the end of the array. Then see which is the number nearest to the end (after the peak) which is also greater than the number just before the peak. The swap the two numbers and then sort all the numbers from peak to the end in ascending order.