Time O(N) , Space O(N)
We do this in single traversal. At each element i, we maintain a current sum. If our curr_sum==k then count++. Also we maintain a map of which sum values have already been found and how many times have they been found. If we have curr_sum- k in our map, it means that we have encountered some subarray with sum=curr_sum-k and if we remove that subarray from the current subarray then we can obtain the sum =k from the remaining part of it. Also, our map has the count of how many times have we encountered such a subarray. We add that count to our count variable. count+=mp[curr_sum-k]