Time O(nlogn) Space O(N)
Instead of going from the given array towards the sorted array, we go the other way round because in the first approach we do not know which position a given element is going to sit in. We make a vector of pairs of element and its position in the given array .Then, we sort the vector (hence the time complexity). Now iterate over the vector and if an element's current position does not match its original position(the second element of each pair), then swap that pair with the pair at the srno of the second element of the ccurrent pair and then increment the count of swap by one.