Time O(N*logK) , Space - O(N)
Given an array of nodes in which each node represents a linked list, we need to merge all the lists into one and return. We use merge sort for this. First using divide part of divide and conquer, we divide the array into groups of 1 or 2 linked lists and then we use our merge procedure to merge each of the pair of linked lists.