Time O(N), Space O(1)
In N nos, if we have some -ve and/or some >n nos then we are sure to find some number from 1 to n that is missing. We check for 1 first then we traverse again and all those elements which are either non-positive or are >n are made to 1. Then traverse again and make the elements at nums[i]th index as -ve.
Then the first positive number in the array after all these operations is the answer. If all the elements are -ve in the array, then the answer will be n+1